ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
scikit-learn==1.4.0
joblib==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.3
pydantic==2.5.3
